refactor: refactor and rebase the code
