+++ app/src/main/java/com/github/se/cyrcle/ui/list/ListScreen.kt
@@ -49,6 +49,7 @@ import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.MutableState
+import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.State
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
@@ -154,7 +155,13 @@ fun SpotListScreen(
       bottomBar = { BottomNavigationBar(navigationActions, selectedItem = Route.LIST) }) {
           innerPadding ->
         Column(modifier = Modifier.fillMaxSize().padding(innerPadding).padding(bottom = 16.dp)) {
-          FilterPanel(parkingViewModel = parkingViewModel, displayHeader = true)
+          FilterPanel(
+              parkingViewModel = parkingViewModel,
+              displayHeader = true,
+              addressViewModel,
+              myLocation,
+              chosenLocation,
+              permissionHandler)
           val listState = rememberLazyListState()
           LazyColumn(state = listState, modifier = Modifier.testTag("SpotListColumn")) {
             // Pinned parking spots if any (includes titles and dividers)
@@ -209,221 +216,6 @@ fun SpotListScreen(
               }
             }
           }
-        }
-      }
-}
-
-@SuppressLint("StateFlowValueCalledInComposition")
-@Composable
-fun FilterHeader(
-    parkingViewModel: ParkingViewModel,
-    addressViewModel: AddressViewModel,
-    myLocation: MutableState<Boolean>,
-    chosenLocation: MutableState<Address>,
-    permissionHandler: PermissionHandler
-) {
-  val showProtectionOptions = remember { mutableStateOf(false) }
-  val showRackTypeOptions = remember { mutableStateOf(false) }
-  val showCapacityOptions = remember { mutableStateOf(false) }
-  var showFilters by remember { mutableStateOf(false) }
-
-  val selectedProtection by parkingViewModel.selectedProtection.collectAsState()
-  val selectedRackTypes by parkingViewModel.selectedRackTypes.collectAsState()
-  val selectedCapacities by parkingViewModel.selectedCapacities.collectAsState()
-  val onlyWithCCTV by parkingViewModel.onlyWithCCTV.collectAsState()
-
-  // Text field visibility
-  val isTextFieldVisible = remember { mutableStateOf(false) }
-
-  // Text field value
-  val textFieldValue = remember { mutableStateOf("") }
-
-  val alpha by animateFloatAsState(targetValue = if (isTextFieldVisible.value) 1f else 0f)
-
-  val radius = parkingViewModel.radius.collectAsState()
-
-  Column(modifier = Modifier.padding(16.dp)) {
-    Row(
-        modifier = Modifier.fillMaxWidth().padding(vertical = 8.dp),
-        horizontalArrangement = Arrangement.End,
-        verticalAlignment = Alignment.CenterVertically) {
-
-          // Button transforms into a close button
-          SmallFloatingActionButton(
-              icon = if (isTextFieldVisible.value) Icons.Default.Close else Icons.Default.Search,
-              onClick = { isTextFieldVisible.value = !isTextFieldVisible.value },
-              modifier = Modifier,
-              contentDescription = "Search List Screen",
-          )
-
-          // Text field slides out to the right of the button
-          if (isTextFieldVisible.value) {
-            SearchBarListScreen(
-                addressViewModel,
-                isTextFieldVisible,
-                myLocation,
-                chosenLocation,
-                permissionHandler,
-                textFieldValue)
-          } else {
-            Text(
-                text = stringResource(R.string.all_parkings_radius, radius.value.toInt()),
-                modifier = Modifier.weight(1f).padding(end = 8.dp),
-                style = MaterialTheme.typography.headlineMedium,
-                color = MaterialTheme.colorScheme.onSurface)
-          }
-
-          SmallFloatingActionButton(
-              onClick = { showFilters = !showFilters },
-              icon = if (showFilters) Icons.Default.Close else Icons.Default.FilterList,
-              contentDescription = "Filter",
-              testTag = "ShowFiltersButton")
-        }
-
-    if (showFilters) {
-      Row(
-          modifier = Modifier.fillMaxWidth().padding(vertical = 8.dp),
-          horizontalArrangement = Arrangement.SpaceBetween) {
-            ClickableText(
-                text = AnnotatedString(stringResource(R.string.list_screen_clear_all)),
-                onClick = { parkingViewModel.clearAllFilterOptions() },
-                style =
-                    MaterialTheme.typography.labelMedium.copy(
-                        color = MaterialTheme.colorScheme.primary),
-                modifier = Modifier.padding(horizontal = 8.dp).testTag("ClearAllFiltersButton"))
-            ClickableText(
-                text = AnnotatedString(stringResource(R.string.list_screen_apply_all)),
-                onClick = { parkingViewModel.selectAllFilterOptions() },
-                style =
-                    MaterialTheme.typography.labelMedium.copy(
-                        color = MaterialTheme.colorScheme.primary),
-                modifier = Modifier.padding(horizontal = 8.dp).testTag("ApplyAllFiltersButton"))
-          }
-
-      // Protection filter
-      FilterSection(
-          title = stringResource(R.string.list_screen_protection),
-          expandedState = showProtectionOptions,
-          onReset = { parkingViewModel.clearProtection() },
-          onApply = { parkingViewModel.selectAllProtection() }) {
-            LazyRow(
-                modifier = Modifier.fillMaxWidth().testTag("ProtectionFilter"),
-                horizontalArrangement = Arrangement.spacedBy(4.dp)) {
-                  items(ParkingProtection.entries.toTypedArray()) { option ->
-                    ToggleButton(
-                        text = option.description,
-                        onClick = { parkingViewModel.toggleProtection(option) },
-                        value = selectedProtection.contains(option),
-                        modifier = Modifier.padding(2.dp),
-                        testTag = "ProtectionFilterItem")
-                  }
-                }
-          }
-
-      // Rack type filter
-      FilterSection(
-          title = stringResource(R.string.list_screen_rack_type),
-          expandedState = showRackTypeOptions,
-          onReset = { parkingViewModel.clearRackType() },
-          onApply = { parkingViewModel.selectAllRackTypes() }) {
-            LazyRow(
-                modifier = Modifier.fillMaxWidth().testTag("RackTypeFilter"),
-                horizontalArrangement = Arrangement.spacedBy(4.dp)) {
-                  items(ParkingRackType.entries.toTypedArray()) { option ->
-                    ToggleButton(
-                        text = option.description,
-                        onClick = { parkingViewModel.toggleRackType(option) },
-                        value = selectedRackTypes.contains(option),
-                        modifier = Modifier.padding(2.dp),
-                        testTag = "RackTypeFilterItem")
-                  }
-                }
-          }
-
-      // Capacity filter
-      FilterSection(
-          title = stringResource(R.string.list_screen_capacity),
-          expandedState = showCapacityOptions,
-          onReset = { parkingViewModel.clearCapacity() },
-          onApply = { parkingViewModel.selectAllCapacities() }) {
-            LazyRow(
-                modifier = Modifier.fillMaxWidth().testTag("CapacityFilter"),
-                horizontalArrangement = Arrangement.spacedBy(4.dp)) {
-                  items(ParkingCapacity.entries.toTypedArray()) { option ->
-                    ToggleButton(
-                        text = option.description,
-                        onClick = { parkingViewModel.toggleCapacity(option) },
-                        value = selectedCapacities.contains(option),
-                        modifier = Modifier.padding(2.dp),
-                        testTag = "CapacityFilterItem")
-                  }
-                }
-          }
-
-      // CCTV filter with checkbox
-      Row(
-          modifier = Modifier.fillMaxWidth().padding(12.dp),
-          verticalAlignment = Alignment.CenterVertically) {
-            Checkbox(
-                checked = onlyWithCCTV,
-                onCheckedChange = { parkingViewModel.setOnlyWithCCTV(it) },
-                modifier = Modifier.testTag("CCTVCheckbox"),
-                colors = getCheckBoxColors(ColorLevel.PRIMARY))
-            Spacer(modifier = Modifier.width(8.dp))
-            Text(
-                text = stringResource(R.string.list_screen_display_only_cctv),
-                style = MaterialTheme.typography.bodyMedium,
-                modifier = Modifier.testTag("CCTVCheckboxLabel"))
-          }
-    }
-  }
-}
-
-@Composable
-fun FilterSection(
-    title: String,
-    expandedState: MutableState<Boolean>,
-    onReset: () -> Unit,
-    onApply: () -> Unit,
-    content: @Composable () -> Unit
-) {
-  Column(
-      modifier =
-          Modifier.padding(vertical = 8.dp)
-              .border(1.dp, Color.Gray, shape = MaterialTheme.shapes.medium)
-              .background(
-                  MaterialTheme.colorScheme.surfaceBright, shape = MaterialTheme.shapes.medium)
-              .padding(8.dp)) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.CenterVertically) {
-              ClickableText(
-                  text = AnnotatedString(stringResource(R.string.list_screen_clear)),
-                  onClick = { onReset() },
-                  style =
-                      MaterialTheme.typography.labelSmall.copy(
-                          color = MaterialTheme.colorScheme.primary),
-                  modifier =
-                      Modifier.padding(horizontal = 4.dp).testTag("Clear" + title + "Button"))
-              Text(
-                  text = title,
-                  style = MaterialTheme.typography.titleMedium,
-                  modifier =
-                      Modifier.clickable(onClick = { expandedState.value = !expandedState.value })
-                          .padding(6.dp),
-                  testTag = title)
-              ClickableText(
-                  text = AnnotatedString(stringResource(R.string.list_screen_apply)),
-                  onClick = { onApply() },
-                  style =
-                      MaterialTheme.typography.labelSmall.copy(
-                          color = MaterialTheme.colorScheme.primary),
-                  modifier = Modifier.padding(horizontal = 4.dp).testTag("Apply${title}Button"))
-            }
-
-        if (expandedState.value) {
-          Box(modifier = Modifier.padding(top = 8.dp)) { content() }
         }
       }
 }
@@ -612,135 +404,4 @@ fun ActionCard(
               )
             }
       }
-}
-
-/**
- * Composable function that displays a search bar for the list screen.
- *
- * @param addressViewModel The ViewModel responsible for managing addresses.
- * @param isTextFieldVisible A state that controls the visibility of the text field.
- * @param myLocation A state that indicates whether the user has selected "My Location".
- * @param chosenLocation A state that holds the address chosen by the user.
- * @param permissionHandler The handler for managing permissions.
- * @param textFieldValue The current value of the text field.
- */
-@SuppressLint("StateFlowValueCalledInComposition")
-@Composable
-fun SearchBarListScreen(
-    addressViewModel: AddressViewModel,
-    isTextFieldVisible: MutableState<Boolean>,
-    myLocation: MutableState<Boolean>,
-    chosenLocation: MutableState<Address>,
-    permissionHandler: PermissionHandler,
-    textFieldValue: MutableState<String>
-) {
-
-  // Initialize FocusManager
-  val focusManager = LocalFocusManager.current
-
-  // Initialize the keyboard controller
-  val virtualKeyboardManager = LocalSoftwareKeyboardController.current
-
-  // List of suggestions from NominatimAPI
-  val listOfSuggestions = addressViewModel.addressList.collectAsState()
-
-  val uniqueSuggestions = remember { mutableStateOf(listOf<Address>()) }
-
-  // Show suggestions screen
-  val showSuggestions = remember { mutableStateOf(false) }
-
-  // value that contain the size fo the textfield to align the suggestion menu below
-  val textFieldSize = remember { mutableStateOf(IntSize.Zero) }
-
-  val alpha by animateFloatAsState(targetValue = if (isTextFieldVisible.value) 1f else 0f)
-
-  // Animation values
-  val slideOffset by
-      animateDpAsState(targetValue = if (isTextFieldVisible.value) 0.dp else (-200).dp)
-
-  Box {
-    OutlinedTextField(
-        value = textFieldValue.value,
-        onValueChange = { textFieldValue.value = it },
-        placeholder = { Text(text = stringResource(R.string.search_bar_placeholder)) },
-        modifier =
-            Modifier.padding(start = 8.dp)
-                .offset(x = slideOffset)
-                .alpha(alpha)
-                .onGloballyPositioned { coordinates -> textFieldSize.value = coordinates.size },
-        colors = getOutlinedTextFieldColorsSearchBar(ColorLevel.PRIMARY),
-        trailingIcon = {
-          if (textFieldValue.value.isNotEmpty()) {
-            Icon(
-                imageVector = Icons.Filled.Clear,
-                contentDescription = "Clear search",
-                tint = defaultOnColor(),
-                modifier =
-                    Modifier.clickable {
-                      textFieldValue.value = ""
-                      showSuggestions.value = false
-                    })
-          }
-        },
-        singleLine = true,
-        keyboardOptions = KeyboardOptions.Default.copy(imeAction = ImeAction.Search),
-        keyboardActions =
-            KeyboardActions(
-                onSearch = {
-                  virtualKeyboardManager?.hide()
-
-                  runBlocking { addressViewModel.search(textFieldValue.value) }
-
-                  showSuggestions.value = true
-                }))
-
-    if (listOfSuggestions.value.size + 1 > 0) {
-      DropdownMenu(
-          expanded = showSuggestions.value,
-          onDismissRequest = { showSuggestions.value = false },
-          modifier =
-              Modifier.width(with(LocalDensity.current) { textFieldSize.value.width.toDp() })) {
-            if (permissionHandler.getLocalisationPerm().value) {
-              DropdownMenuItem(
-                  text = { Text("My Location") },
-                  onClick = {
-                    myLocation.value = true
-                    showSuggestions.value = false
-                    textFieldValue.value = "My Location"
-                    isTextFieldVisible.value = false
-                    focusManager.clearFocus()
-                  })
-            }
-
-            val seenNames = mutableSetOf<String>()
-            uniqueSuggestions.value =
-                listOfSuggestions.value.filter { suggestion ->
-                  val displayName =
-                      suggestion.suggestionFormatDisplayName(maxSuggestionDisplayNameLengthList)
-                  if (displayName in seenNames) {
-                    false
-                  } else {
-                    seenNames.add(displayName)
-                    true
-                  }
-                }
-
-            for (address in uniqueSuggestions.value.take(6)) {
-              DropdownMenuItem(
-                  text = {
-                    Text(address.suggestionFormatDisplayName(maxSuggestionDisplayNameLengthList))
-                  },
-                  onClick = {
-                    chosenLocation.value = address
-                    showSuggestions.value = false
-                    myLocation.value = false
-                    isTextFieldVisible.value = false
-                    textFieldValue.value =
-                        address.suggestionFormatDisplayName(maxSuggestionDisplayNameLengthList)
-                    focusManager.clearFocus()
-                  })
-            }
-          }
-    }
-  }
 }
